feat(): refactor Combobox stories
